Upon wizard submission, avoid reloading the whole app
This full reload was a workaround for an earlier implementation of the data flow, but I don't see any point in it now. Avoiding the reload should also improve the JS coverage metrics, since the reload appears to have been resetting the coverage stats and resulted in no coverage data for the assignment wizard components and related code.